split connection/mod into sync and async
